backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic
langchain
langchain-community
langchain-openai
faiss-cpu
rank-bm25
pypdf
